Move getting recycled bitmaps into stream resizer
Lets us actually determine the size of the bitmap
we want to load and use that to see if a recycled
one is available. Otherwise we would have to pass
in values which would be difficult to keep track
of. Side affect is we need to get two input
different input streams from the disk cache, one
to check the size and one to actually load the
image.