BaseImageLoader manages refs and loader lifecycle
Allows simpler callbacks in the presenter or
higher levels and moves responsibility for not 
leaking referneces into the image loader. ALso
combines ImageLoader and AssetPathConverter 
interface to simplify the external interface and
make it possible to use a presenter with a model
that already has a path or otherwise doesn't need
to fetch one.