Don't make subclasses of loaders use internal cb
Simplifies subclasses and makes it easier to 
make callbacks leak proof without sacrificing
safety. If the subclasses have a strong reference
only to the internal callback, they will not 
retain a strong reference to the presenter or the
image view since the internal callback only has a
weak reference to a callback generated by the 
image presenter.