Replace PathLoader with ModelStreamLoader
A major change that should allow a larger range
of images to be used (i.e. those to which it is
only possible to obtain an InputStream and not
a path) as well as making it possible to download
and store images directly in the cache without
an additional class or disk cache.