Create a simple static interface for Glide
Allows most of the functionality of the library
without having to directly create and manage an
ImageManager and ImagePresenters.

Doesn't currently allow for all functionality and
is somewhat less efficient because it encourages
allocations of unneeded objects, but it is easier
to use and should cover enough cases to be useful.