Allow match_parent/wrap_content view dimensions
By default if the view has wrap_content for either
the width or the height (or both), we will load
the image in at the original size. Trying to apply
either FitCenter or CenterCrop to the load will
still result in an exception since we would need
concrete sizes to transform the image.