Add a Size + Config Strategy.
Fixes #335.

Allows us to re-use Bitmaps with configs other
than ARGB_8888. Differs from AttributeStrategy in
that it sorts Bitmaps by config only, and allows
Bitmaps with a given config to be reconfigured to
change their dimensions. In addition, we allow
switching configs between the hidden GIF config
and ARGB_8888 which appears not to trigger #301.