Allow removing size callbacks from Targets.
There's a race where a Target can be cleared before a View is laid out. If that happens and the view is laid out later on, we will crash. Removing the callback from the ViewTarget when the Target is cleared resolves that issue.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=155384032