Use accurate rounding in FitCenter when checking to see if we can skip
floor is necessary to make sure that we populate all pixels in the
Bitmap, but it's not necessary to use to check if our input Bitmap has
already been transformed and can be returned as is. Using floor can
cause us to re-draw a Bitmap unnecessarily.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=167789795