Add side by side script to upload to bintray.
This allows us to test uploading to bintray without eliminating our
ability to upload to sonatype and maven central directly. If uploading
to bintray seems to work for a version or two and we figure out syncing
to maven central and snapshots, we can later remove the
maven_central.gradle script entirely.