Fail CachingTest if our weak reference is never cleared.
We're waiting for a specific condition that should occur, but it's
dependent on the VM. I haven't seen it fail with 100 attempts, but it's
possible it might. Failing with a specific assertion should make it
easier to distinguish between failed tests and flakes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177625844