Only avoid restarting old requests when memory cache is disabled, not just set
You can set skipMemoryCache to true or false, or you can not set it at
all. Previously we were avoiding using the previous request if
skipMemoryCache is set, which means that not setting skipMemoryCache
behaves differently than setting skipMemoryCache to false. That's not
necessarily a bug, but it's not expected either. Now we only avoid using
the previous request if skipMemoryCache is explicitly set to true.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178704109