Add a Gradle task to regenerate test resources for compiler tests.
Doing so requires the cooperation of the tests themselves by relying
on a JUnit rule to overwrite the test resources with the current output
of the annotation processor if a specific property is set. 

This is a bit of hack, but it’s at least much nicer than manually 
copy/pasting a bunch of file contents around every time we add or
re-order annotations.