Include the debug aar in release artifacts for Android projects.
We removed the release variant a while ago to speed up the build, which
has the side affect of removing the release aar from artifacts. Since
we expect the debug and release variants to be identical (hence why
we disabled the release variant), it should be safe to just use the 
debug aar instead. We will have to specify it explicitly since android’s
rules unsurprisingly only add the release variant by default.