Fall back to decoding resources from other packages directly.
AppCompat tries to use internal resources to determine whether or not it's capable of decoding vector graphics. The ids of those resources will change across different versions of the support library. If the versions of the support library in the two apps don't match, there might be an id mismatch, which leads to an error. In some cases we can decode the resource anyway by bypassing the support library entirely.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190825442