Remove IntRange in RequestOptions.overrideOf.
It’s not possible to whitelist Target.SIZE_ORIGINAL, so the range doesn’t make sense. A long term answer might be to use SIZE_ORIGINAL to be Int.MAX_VALUE instead, but in the short term the annotation is incorrect and not present in the non-static versions of the method.

Fixes #3443.