Remove failing ActiveResourcesTests.
These tests can't reliably pass because they rely on the behavior of
WeakReferences which are documented as being inconsistent. In particular
we can't really simulate an item being added to the reference queue, but
not yet cleared by the system. We can simulate the end state, but not
the intermediate state.

PiperOrigin-RevId: 250577546