Add Copybara migrations to convert to androidx
The external version of Glide will support androidx after https://github.com/bumptech/glide/pull/3518, but Google only supports a small subset of all the androidx packages. As a result we need to migrate back and forth to keep the external version running on androidx and the internal version running on support. As they migrate google3 to androidx, we can remove these transformations.

PiperOrigin-RevId: 250593532