Extend raw ComparableSubject instead of supplying type parameters.
The self-type parameter is being removed from ComparableSubject. In order to remove one type parameter while leaving the other in place, it's necessary either to make an atomic change (which is tricky for third-party code) or to temporarily remove *both* type parameters and then put the one back. This CL implements the latter. This CL will temporarily produce rawtypes warnings -- and temporarily eliminate type-checking of calls to the ComparableSubject methods -- which will go away when I remove the type parameters (as soon as this batch of CLs is submitted).

This CL requires Truth 0.45. I should already have another CL out to you (perhaps even already submitted) to make other changes to prepare for Truth 0.45.

PiperOrigin-RevId: 251895062