Improve comments and consistency of sampling in DownsampleStrategies.
FitCenter, CenterInside, and CenterOutside now all reliably prefer 
quality on versions < KitKat where we can only do power of two 
downsampling.

I’ve also fixed a bug in FitCenter and CenterInside where we would 
downsample to 2x the requested size in some cases where there was
an exact power of 2 match in one dimension but not the other. The logic
to fix the bug isn’t perfect, but it keeps this change small and avoids
making any API changes.