Create annotation glide module classes lazily but outside a lock.
This means we might instantiate GlideModule classes multiple times if
multiple threads race and call Glide.with(). Only one instance will
actually be used even if this happens.

Instantiating modules outside of a lock allows to both:
1. Provide an opportunity for dependency injection into GlideModule
constructors without introducing deadlock.
2. Lazily create modules.

We could instantiate the modules exactly once if we initialize them
eagerly, but that would slow down application startup. We can
instantiate modules exactly once, but can introduce deadlock if
dependency injection is used (Glide.get() acquires the Glide lock, then
the DI lock. DI.get() acquires the DI lock, then the Glide lock).

The primary cost of instantiating (and injecting) glide modules is
probably in the reflection and class loading. Most of that is already
memoized, so it's not super expensive to do it more than once. Most
applications won't race anyway, so most of the time the cost is
effectively non-existent (we hope).

A future change will allow passing in a Context to the constructor
of modules to allow injection in the constructors only.

PiperOrigin-RevId: 253114941