Allow AppGlideModules to optionally accept a Context in their constructor.
The Context can then be used to safely inject parameters and/or classes
outside of Glide's initialization lock. The initialization will happen
exactly once per AppGlideModule instance, but the lack of a lock means
that more than one AppGlideModule instance may be created.

We can't both not use a lock and guarantee that initialization will
occur exactly once. Rare race conditions causing double initialization
are probably worth the cost, especially because most DI frameworks will
memoize and minimize the cost of the second injection if it does occur.

PiperOrigin-RevId: 253302453