Add a content provider to print some Glide stats to bug reports.
Right now this just prints memory sizes. The implementations are a bit
of a hack right now, but should be mostly tolerable in that very little
additional data is exposed. I didn't want to modify the interfaces
because handling custom implementations doesn't seem worth breaking
existing implementations. We can do so in the future if this functionality turns out to be useful.

PiperOrigin-RevId: 258997344