Add support for P3 and SRGB in Glide.
Allows users of the library to manually set an option to support displaying wide gamut images with the DISPLAY_P3 color space on P+.  Defaults Glide to using SRGB on O and for all loads where the option isn't set to DISPLAY_P3.

Does not allow callers to specify other color spaces. Does not allow callers to convert SRGB images to DISPLAY_P3. Does not verify that screens/renderers are capable of displaying P3 images when it's requested.

PiperOrigin-RevId: 260580611