Avoid deadlock in RequestCoordinators by manually tracking child Request states.
This is an unfortunately large change and isn't complete because it's not strictly thread safe. I expect that it's close to correct because typically the state of each sub request will only be modified by a single thread at a time, but there's no strict guarantee. More locking is needed in a follow up. I believe we'll be able to lock when checking/changing state, but avoid locking when calling in to sub requests to avoid the deadlock.

PiperOrigin-RevId: 260960463