Use sourceSets instead of JavaCompileProvider to generate javadoc.
JavaCompileProvider's source list is always empty.

This change also removes the public jar generation and replaces it with
a single fat jar that's meant for internal use only. As part of those
changes I also removed unused logic that allowed a variant to be
specified. Glide only supports the "debug" variant to eliminate the
redundant compilation/unit tests associated with the release variant.